requirements.txt
beautifulsoup4==4.10.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
future==0.18.2
google==3.0.0
google-api-core==2.5.0
google-api-python-client==2.36.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
httplib2==0.20.4
idna==3.3
oauthlib==3.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dotenv==0.19.2
PyVimeo==1.1.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
soupsieve==2.3.1
tinydb==4.6.1
tuspy==0.2.5
uritemplate==4.1.1
urllib3==1.26.8
